Salvar agora funciona, falta ajeitar o gera prova para mostrar conteudos da disciplina selecionada.

+++ ProjetoMavenJSFPrimeFaces/src/main/java/sistema/dao/ConteudoDAO.java
@@ -3,6 +3,6 @@ package sistema.dao;
 import sistema.dao.generic.DAOImpl;
 import sistema.modelos.Conteudo;
 
-public class ConteudoDAO extends DAOImpl<Conteudo, Integer>{
+public class ConteudoDAO extends DAOImpl<Conteudo, Long>{
 
 }
