+++ ProjetoMavenJSFPrimeFaces/src/main/java/sistema/dao/DisciplinaDAO.java
@@ -2,5 +2,5 @@ package sistema.dao;
 import sistema.dao.generic.DAOImpl;
 import sistema.modelos.Disciplina;
 
-public class DisciplinaDAO extends DAOImpl<Disciplina, Integer>{
+public class DisciplinaDAO extends DAOImpl<Disciplina, Long>{
 }
